DBZ-776 Extract metrics to a common code
